attribution_worker/pyproject.toml
[project]
name = "infini-gram-attribution-worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings==2.3.4",
    "pydantic==2.10.6",
    "saq[postgres]==0.22.4",
    "psycopg[binary,pool]>=3.2.6",
    "transformers==4.49.0",
    "infini-gram-processor",
    "numpy<2.0.0",
]
